+++ src/main/java/handler/Handler.java
@@ -15,7 +15,9 @@ import com.amazonaws.services.s3.model.S3Object;
 import com.amazonaws.services.s3.model.S3ObjectInputStream;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import parser.MailParser;
+import parser.MailParserFactory;
 import parser.PscDiscrepancySurveyCsvProcessor;
+import parser.PscDiscrepancySurveyCsvProcessorFactory;
 import service.AmazonS3Service;
 import uk.gov.companieshouse.environment.EnvironmentReader;
 import uk.gov.companieshouse.environment.impl.EnvironmentReaderImpl;
