+++ src/main/java/parser/PscDiscrepancySurveyCsvProcessorFactory.java
@@ -1,10 +1,10 @@
-package handler;
+package parser;
 
-import parser.PscDiscrepancySurveyCsvProcessor;
+import parser.PscDiscrepancySurveyCsvProcessor.PscDiscrepancyCreatedListener;
 
 public class PscDiscrepancySurveyCsvProcessorFactory {
     public PscDiscrepancySurveyCsvProcessor createPscDiscrepancySurveyCsvProcessor(
-                    byte[] extractedCsv, PscDiscrepancyFoundListenerImpl listener) {
+                    byte[] extractedCsv, PscDiscrepancyCreatedListener listener) {
         return new PscDiscrepancySurveyCsvProcessor(extractedCsv, listener);
     }
 }
